requirements.txt
Django==2.2
djangorestframework==3.10.3
django-countries==6.1.3
